Move logging flags init to from run.go to root.go
