+++ sample/peer/cmd/run.go
@@ -22,7 +22,6 @@ import (
 	"strconv"
 
 	"github.com/a8m/envsubst"
-	logging "github.com/op/go-logging"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"google.golang.org/grpc"
@@ -73,14 +72,6 @@ func init() {
 		defUsigEnclaveFile, "USIG enclave file")
 	must(viper.BindPFlag("usig.enclaveFile",
 		runCmd.Flags().Lookup("usig-enclave-file")))
-
-	rootCmd.PersistentFlags().String("logging-level", "", "logging level")
-	must(viper.BindPFlag("logging.level",
-		rootCmd.PersistentFlags().Lookup("logging-level")))
-
-	rootCmd.PersistentFlags().String("logging-file", "", "logging file")
-	must(viper.BindPFlag("logging.file",
-		rootCmd.PersistentFlags().Lookup("logging-file")))
 }
 
 type replicaStack struct {
@@ -158,26 +149,4 @@ func run() error {
 	}()
 
 	return <-srvErrChan
-}
-
-func getLoggingOptions() ([]logger.Option, error) {
-	opts := []logger.Option{}
-
-	if viper.GetString("logging.level") != "" {
-		logLevel, err := logging.LogLevel(viper.GetString("logging.level"))
-		if err != nil {
-			return nil, fmt.Errorf("failed to set logging level: %s", err)
-		}
-		opts = append(opts, logger.WithLogLevel(logLevel))
-	}
-
-	if viper.GetString("logging.file") != "" {
-		logFile, err := os.OpenFile(viper.GetString("logging.file"), os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0600)
-		if err != nil {
-			return nil, fmt.Errorf("failed to open logging file: %s", err)
-		}
-		opts = append(opts, logger.WithLogFile(logFile))
-	}
-
-	return opts, nil
 }
